+++ mayan/apps/metadata/tests/literals.py
@@ -13,10 +13,8 @@ TEST_METADATA_TYPES_FIXTURES = [
     {'name': 'test_metadata_type_1', 'label': 'test metadata type 1'},
     {'name': 'test_metadata_type_2', 'label': 'test metadata type 2'},
 ]
-TEST_METADATA_TYPE_LABEL = 'test metadata type'
 TEST_METADATA_TYPE_LABEL_2 = 'test metadata type label 2'
 TEST_METADATA_TYPE_LABEL_EDITED = 'test metadata type label edited'
-TEST_METADATA_TYPE_NAME = 'test'
 TEST_METADATA_TYPE_NAME_2 = 'test metadata type name 2'
 TEST_METADATA_TYPE_NAME_EDITED = 'test metadata type name edited'
 TEST_METADATA_VALUE = 'test value'